primeira correção de cotação
ainda em endamento, será necessária nova correção no esquema relacional

+++ src/main/java/br/com/synergy/repository/CotacoesMateriais.java
@@ -29,7 +29,7 @@ public class CotacoesMateriais implements Serializable {
 		em.remove(cotacao);
 	}
 	
-	public CotacaoMaterial buscaPorId(Integer id){
+	public CotacaoMaterial buscaPorId(Long id){
 		return em.find(CotacaoMaterial.class, id);
 	}
